Poprawki

+++ kierownik.c
@@ -1,5 +1,6 @@
 #include "common.h"
 #include <sys/wait.h>
+#include <time.h>
 
 static void ewakuacja(int msgid, SharedState *stan) {
     stan->ewakuacja_trwa = 1;
@@ -7,20 +8,29 @@ static void ewakuacja(int msgid, SharedState *stan) {
 
     for (int i = 0; i < LICZBA_SEKTOROW; i++) {
         MsgSterujacy msg = {10 + i, 3, i};
-        msgsnd(msgid, &msg, sizeof(int) * 2, 0);
+        if (msgsnd(msgid, &msg, sizeof(int) * 2, 0) == -1) perror("msgsnd");
     }
+
+    printf("[KIEROWNIK] EWAKUACJA\n");
+    fflush(stdout);
 
     int raporty = 0;
     while (raporty < LICZBA_SEKTOROW) {
         MsgSterujacy rap;
-        if (msgrcv(msgid, &rap, sizeof(int) * 2, 99, 0) >= 0) {
+        ssize_t res = msgrcv(msgid, &rap, sizeof(int) * 2, 99, 0);
+        if (res >= 0) {
             printf("[RAPORT] Sektor %d pusty\n", rap.sektor_id);
             fflush(stdout);
             raporty++;
+        } else {
+            if (errno == EIDRM || errno == EINVAL) break;
+            if (errno == EINTR) continue;
+            perror("msgrcv");
+            break;
         }
     }
 
-    printf("[KIEROWNIK] Koniec\n");
+    printf("[KIEROWNIK] Koniec symulacji\n");
     fflush(stdout);
 }
 
@@ -28,13 +38,13 @@ int main() {
     setbuf(stdout, NULL);
 
     int msgid = msgget(KEY_MSG, 0600);
-    if (msgid == -1) exit(1);
+    if (msgid == -1) { perror("msgget"); exit(1); }
 
     int shmid = shmget(KEY_SHM, sizeof(SharedState), 0600);
-    if (shmid == -1) exit(1);
+    if (shmid == -1) { perror("shmget"); exit(1); }
 
     SharedState *stan = (SharedState*)shmat(shmid, NULL, 0);
-    if (stan == (void*)-1) exit(1);
+    if (stan == (void*)-1) { perror("shmat"); exit(1); }
 
     stan->status_meczu = 0;
     stan->czas_pozostaly = CZAS_PRZED_MECZEM;
@@ -65,8 +75,11 @@ int main() {
         exit(0);
     }
 
-    printf("Komendy: 1-stop sektora, 2-start sektora, 3-ewakuacja\n");
+    printf("Komendy: 1-stop, 2-start, 3-ewakuacja\n");
     fflush(stdout);
+
+    fd_set readfds;
+    struct timeval tv;
 
     while (1) {
         if (waitpid(zegar_pid, NULL, WNOHANG) > 0) {
@@ -74,27 +87,36 @@ int main() {
             break;
         }
 
-        int cmd;
-        if (scanf("%d", &cmd) == 1) {
-            if (cmd == 3) {
-                kill(zegar_pid, SIGTERM);
-                waitpid(zegar_pid, NULL, 0);
-                ewakuacja(msgid, stan);
-                break;
-            }
-            if (cmd == 1 || cmd == 2) {
-                int s;
-                printf("Sektor (0-7): ");
-                fflush(stdout);
-                if (scanf("%d", &s) == 1 && s >= 0 && s < LICZBA_SEKTOROW) {
-                    MsgSterujacy msg = {10 + s, cmd, s};
-                    msgsnd(msgid, &msg, sizeof(int) * 2, 0);
+        FD_ZERO(&readfds);
+        FD_SET(STDIN_FILENO, &readfds);
+        tv.tv_sec = 0;
+        tv.tv_usec = 500000;
+
+        int ret = select(STDIN_FILENO + 1, &readfds, NULL, NULL, &tv);
+        if (ret > 0) {
+            int cmd;
+            if (scanf("%d", &cmd) == 1) {
+                if (cmd == 3) {
+                    kill(zegar_pid, SIGTERM);
+                    waitpid(zegar_pid, NULL, 0);
+                    ewakuacja(msgid, stan);
+                    break;
+                }
+                if (cmd == 1 || cmd == 2) {
+                    int s;
+                    printf("Sektor (0-7): ");
+                    fflush(stdout);
+                    if (scanf("%d", &s) == 1 && s >= 0 && s < LICZBA_SEKTOROW) {
+                        MsgSterujacy msg = {10 + s, cmd, s};
+                        if (msgsnd(msgid, &msg, sizeof(int) * 2, 0) == -1) perror("msgsnd");
+                    }
                 }
             }
         }
-        usleep(100000);
     }
 
     shmdt(stan);
+    kill(zegar_pid, SIGTERM);
+    waitpid(zegar_pid, NULL, 0);
     return 0;
 }
